Add a CPIO_newc convenience method to make symlinks
The main issue with CPIOSymlink is that it follows CPIOMember
convention, i.e. the name of the link is the first argument and
the target of the link follows. Unfortunately, this is exactly
the reversed order from the command-line "ln" utility, causing
some confusion.

I don't want to reverse the parameters to the CPIOSymlink
constructor, because that might be even more confusing...

Instead, let's add a convenience method to the CPIO_newc class,
which will be (hopefully) used by most people. Those who need to
dig into the gory details of the underlying constructor will
probably understand the logic behind the reversed order...

Signed-off-by: Petr Tesarik <ptesarik@suse.com>