Shared object dependencies
After toying with the idea of parsing ELF files, this is now
implemented by calling "ldd" on the binary. Parsing the ELF file
is not so difficult, but finding the corresponding file path is.
That's mainly because of some advanced features of the linker,
such as glibc tunables and hwcaps.

The class is now put into install.{cc,h}, because it is closely
related to installing binaries in the initrd.

Signed-off-by: Petr Tesarik <ptesarik@suse.com>